+++ cdx_core/config/routes.rb
@@ -51,11 +51,7 @@ Rails.application.routes.draw do
 
   get 'settings' => 'home#settings'
 
-  resources :test_orders_state, only: [:index, :show] do
-    collection do
-      get :patients
-    end
-  end
+  resources :test_orders_state, only: [:index, :show]
 
   resources :encounters do
     resource :patient_results, only: [:update] do
