+++ cdx_vietnam/app/middleware/patients/list_csv.rb
