+++ cdx_vietnam/config/routes.rb
@@ -4,4 +4,10 @@ Rails.application.routes.draw do
       get 'retry'
     end
   end
+
+  resources :test_orders_state, only: [:index, :show] do
+    collection do
+      get :patients
+    end
+  end
 end
